feat: make function tests easier to devlop by seperating them into their own test file
